% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xgboost.R
\name{convert_xgboost_rpf}
\alias{convert_xgboost_rpf}
\title{Convert xgboost to rpf object}
\usage{
convert_xgboost_rpf(xg, x, y)
}
\arguments{
\item{xg}{xgboost object}

\item{x}{data used to train the xgboost model}

\item{y}{target used to train the xgboost model}
}
\value{
rpf object
}
\description{
Convert xgboost to rpf object
}
