% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{extract_components}
\alias{extract_components}
\title{Extract predicted components from a Random Planted Forest}
\usage{
extract_components(object, new_data, predictors = NULL)
}
\arguments{
\item{object}{A fit object of class \code{\link{rpf}}.}

\item{new_data}{Data for new observations to predict.}

\item{predictors}{}
}
\value{
A numeric \code{matrix}
}
\description{
Extract predicted components from a Random Planted Forest
}
\examples{

# Regression task, only numeric predictors
train <-  mtcars[1:20, ]
test <-  mtcars[21:32, ]

set.seed(23)
object <- rpf(mpg ~ ., data = train, max_interaction = 3, ntrees = 35)

# One component, first predictor column in test data
extract_components_by_name(object, test, predictors = c("cyl"))

# Two components, 1st and 3rd predictor columns in test data
extract_components_by_name(object, test, predictors = c("cyl", "hp"))

# trying to pass all columns in test data errors since outcome "mpg" is included
extract_components_by_name(object, test, predictors = names(test))

# If `predictors` is not specified, we don't know what to do. Are you looking for `[predict.rpf]`?
extract_components_by_name(object, test)
}
