% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_factor_by_response}
\alias{order_factor_by_response}
\title{Order factor levels by response}
\usage{
order_factor_by_response(x, y)
}
\arguments{
\item{x}{Factor variable to order}

\item{y}{Response}
}
\value{
Re-ordered (ordered) factor
}
\description{
Regression: Order by mean(y)
Binary classification: Order by mean(y==1)
Multiclass: Order by first principal component of the weighted covariance
matrix of the contingency table
}
\details{
See https://doi.org/10.7717/peerj.6339 for details.
}
