% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pca_order}
\alias{pca_order}
\title{Order factor levels by first principal component of the weighted covariance matrix of the contingency table}
\usage{
pca_order(x, y)
}
\arguments{
\item{x}{Factor variable to order}

\item{y}{Response}
}
\value{
Order of factor levels
}
\description{
Reference: Coppersmith, D., Hong, S.J. & Hosking, J.R. (1999) Partitioning Nominal Attributes in Decision Trees. Data Min Knowl Discov 3:197. \doi{10.1023/A:1009869804967}.
}
