% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot_main_effect}
\alias{plot_main_effect}
\alias{plot_twoway_effects}
\alias{plot_threeway_effects}
\alias{autoplot.rpf_components}
\title{Plot Prediction Components}
\usage{
plot_main_effect(components, predictor, ...)

plot_twoway_effects(components, predictors, ...)

plot_threeway_effects(components, predictors, ...)

\method{autoplot}{rpf_components}(object, predictors, ...)
}
\arguments{
\item{components, object}{Predicted components of an \code{rpf()} including the original data the model was fit on, as
returned by \code{predict_components()}}

\item{predictor, predictors}{\verb{[character]} vector of predictor names, e.g. \code{"x1"} to plot main effect of \code{x1}, and
\code{c("x1", "x2")} to plot the interaction term \code{x1:x2}.}

\item{...}{Unused}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Prediction Components
}
\examples{

# introduce factor variables to show categorical feature handling
mtcars$cyl <- factor(mtcars$cyl)
mtcars$vs <- factor(mtcars$vs)

# Fit forest, extract components
set.seed(12)
rpfit <- rpf(mpg ~ cyl + wt + hp + drat + vs, data = mtcars, ntrees = 25, max_interaction = 3)
components <- predict_components(rpfit, mtcars)

# Main effects ----
plot_main_effect(components, "wt")
plot_main_effect(components, "drat")
plot_main_effect(components, "cyl")

# 2-degree interaction effects ----
# 2d continuous, scatterplot of arbitrary orientation
plot_twoway_effects(components, c("wt", "drat"))
plot_twoway_effects(components, c("drat", "wt"))

# continuous + categorical (forces continuous on x axis, colors by categorical)
plot_twoway_effects(components, c("wt", "cyl"))
# identical: plot_twoway_effects(components, c("cyl", "wt"))

# 2d categorical, heatmap of arbitrary orientation
plot_twoway_effects(components, c("vs", "cyl"))
plot_twoway_effects(components, c("cyl", "vs"))
# plot_threeway_effects(components, c("hr", "temp", "workingday"))
}
