% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rpf.R
\name{predict_rpf_bridge}
\alias{predict_rpf_bridge}
\title{Bridge: Passes new data to corresponding predict function}
\usage{
predict_rpf_bridge(type, object, predictors, ...)
}
\description{
Bridge: Passes new data to corresponding predict function
}
