% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rpf.R
\name{predict_rpf_class}
\alias{predict_rpf_class}
\title{Predict function for classification: Class prediction}
\usage{
predict_rpf_class(object, new_data, components, ...)
}
\description{
Predict function for classification: Class prediction
}
