% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rpf.R
\name{predict_rpf_numeric}
\alias{predict_rpf_numeric}
\title{Predict function for numeric outcome / regression}
\usage{
predict_rpf_numeric(object, new_data, components, ...)
}
\description{
Predict function for numeric outcome / regression
}
