% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rpf.R
\name{predict_rpf_prob}
\alias{predict_rpf_prob}
\title{Predict function for classification: Probability prediction}
\usage{
predict_rpf_prob(object, new_data, components, ...)
}
\description{
Predict function for classification: Probability prediction
}
