% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rpf_forest}
\alias{print.rpf_forest}
\alias{str.rpf_forest}
\title{Compact printing of forest structures}
\usage{
\method{print}{rpf_forest}(x, ...)

\method{str}{rpf_forest}(object, ...)
}
\arguments{
\item{x}{\verb{[rpf_forest]}: Flattened forest, as in \verb{$forest} of an \code{\link{rpf}}.}

\item{...}{Further arguments passed to or from other methods.}

\item{object}{\verb{[rpf_forest]}: Flattened forest.}
}
\description{
These methods are provided to avoid flooding the console with long nested lists containing tree structures.
}
\examples{

rpfit <- rpf(mpg ~ cyl + wt, data = mtcars, ntrees = 10)
print(rpfit$forest)
str(rpfit$forest)
}
\seealso{
\code{\link{rpf}}
}
