% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{extract_components}
\alias{extract_components}
\title{Extract predicted components from a Random Planted Forest}
\usage{
extract_components(object, new_data, predictors = NULL)
}
\arguments{
\item{object}{A fit object of class \code{\link{rpf}}.}

\item{new_data}{Data for new observations to predict.}

\item{predictors}{\code{\link{character}} Vector of one or more column names of predictor variables in \code{new_data}
to extract components for.}
}
\value{
A numeric \code{matrix}
}
\description{
Extract predicted components from a Random Planted Forest
}
\examples{

# Regression task, only numeric predictors
train <-  mtcars[1:20, ]
test <-  mtcars[21:32, ]

set.seed(23)
object <- rpf(mpg ~ ., data = train, max_interaction = 3, ntrees = 30)

# One component, first predictor column in test data
extract_components(object, test, predictors = c("cyl"))

# Two components, 1st and 3rd predictor columns in test data
extract_components(object, test, predictors = c("cyl", "hp"))

\dontrun{
# trying to pass all columns in test data errors since target "mpg" is included
extract_components(object, test, predictors = names(test))

# passing all predictor names by dropping `"mpg"` first
extract_components(object, test, predictors = names(test[, -1]))

# If `predictors` is not specified, we don't know what to do
extract_components(object, test)
}
}
