% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rpf.R
\name{predict.Rcpp_RandomPlantedForest}
\alias{predict.Rcpp_RandomPlantedForest}
\title{Random Planted Forest Predictions}
\usage{
\method{predict}{Rcpp_RandomPlantedForest}(object, new_data, components, ...)
}
\arguments{
\item{object}{A fit object as returned by \code{\link[=rpf]{rpf()}}.}

\item{new_data}{Matrix or vector input for new observations.}

\item{components}{TODO}

\item{...}{Unused.}
}
\value{
Vector of predictions (FIXME: Type, dimension, response vs. prob)
}
\description{
Random Planted Forest Predictions
}
\examples{
\dontrun{
# This doesn't work just yet
 rpfit <- rpf(mtcars$mpg, as.matrix(mtcars[, c("cyl", "wt")]))
 predict(rpfit, mtcars[1:5, c("cyl", "wt")], components = 1)
}
}
