% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rpf.R
\name{predict.rpf}
\alias{predict.rpf}
\title{Random Planted Forest Predictions}
\usage{
\method{predict}{rpf}(object, new_data, type = "numeric", components = 0, ...)
}
\arguments{
\item{object}{A fit object as returned by \code{\link[=rpf]{rpf()}}.}

\item{new_data}{Input for new observations.}

\item{type}{\verb{["numeric"]}: One of \code{"numeric"} for regression outcomes,
\code{"class"} for class predictions or \code{"prob"} for probability predictions.}

\item{components}{\verb{[0]} TODO.}

\item{...}{Unused.}
}
\value{
For regression: A data.frame with column \code{.pred} with the same
number of rows as \code{new_data}.
For Classification: TBI
}
\description{
Random Planted Forest Predictions
}
\examples{
\dontrun{
rpfit <- rpf(y = mtcars$mpg, x = as.matrix(mtcars[, c("cyl", "wt")]))
predict(rpfit, mtcars[, c("cyl", "wt")], components = 0)
}
}
