% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rpf}
\alias{print.rpf}
\title{Print an rpf}
\usage{
\method{print}{rpf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rpf}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly: \code{x}.
}
\description{
Print an rpf
}
\examples{
rpf(mpg ~ cyl + wt + drat, data = mtcars, max_interaction = 2, ntrees = 10)
}
\seealso{
\code{\link{rpf}}.
}
