% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purify.R
\name{purify}
\alias{purify}
\alias{purify.default}
\alias{purify.rpf}
\alias{is_purified}
\title{Purify a Random Planted Forest}
\usage{
purify(x, ...)

\method{purify}{default}(x, ...)

\method{purify}{rpf}(x, ...)

is_purified(x)
}
\arguments{
\item{x}{And object of class \code{rpf}.}

\item{...}{(Unused)}
}
\value{
Invisibly: The \code{\link{rpf}} object.
}
\description{
TODO: Explain what this does
}
\details{
Unless \code{\link[=rpf]{rpf()}} is called with \code{purify = TRUE}, the forest has to be purified after fit
to ensure the components extracted by \code{\link[=predict_components]{predict_components()}} are valid.
\code{\link[=predict_components]{predict_components()}} will automatically purify a forest if \code{\link[=is_purified]{is_purified()}} reports \code{FALSE}.
}
\examples{
rpfit <- rpf(mpg ~., data = mtcars, max_interaction = 2, ntrees = 10)
purify(rpfit)
}
