% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomPlantedForest-package.R
\docType{package}
\name{randomPlantedForest-package}
\alias{randomPlantedForest}
\alias{randomPlantedForest-package}
\title{randomPlantedForest: Random Planted Forest: A Directly Interpretable Tree Ensemble}
\description{
Implements Random Planted Forest (Hiabu, Mammen and Meyer, 2020) \doi{10.48550/arXiv.2012.14563}, a directly interpretable tree ensemble for regression and classification. Trees are restricted to a maximum number of interacting predictors, so the fitted model decomposes exactly into main effects and low-order interactions, a functional ANOVA decomposition that can be extracted for interpretation. Fitting and prediction are multithreaded, and fitted models can be serialized.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PlantedML/randomPlantedForest}
  \item \url{https://plantedml.com/randomPlantedForest/}
  \item Report bugs at \url{https://github.com/PlantedML/randomPlantedForest/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{burk@leibniz-bips.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

Authors:
\itemize{
  \item Lukas Burk \email{burk@leibniz-bips.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})
  \item Joseph Theo Meyer
  \item Munir Hiabu (\href{https://orcid.org/0000-0001-5846-667X}{ORCID})
  \item Maike Spankus
  \item Marvin N. Wright \email{cran@wrig.de} (\href{https://orcid.org/0000-0002-8542-6291}{ORCID})
}

}
\keyword{internal}
