% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomPlantedForest-package.R
\docType{package}
\name{randomPlantedForest-package}
\alias{randomPlantedForest}
\alias{randomPlantedForest-package}
\title{randomPlantedForest: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PlantedML/randomPlantedForest}
  \item Report bugs at \url{https://github.com/PlantedML/randomPlantedForest/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{github@quantenbrot.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID}) [contributor]

Authors:
\itemize{
  \item Joseph Theo Meyer
  \item Munir Hiabu
  \item Maike
  \item Marvin N. Wright \email{cran@wrig.de}
}

}
\keyword{internal}
