% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomPlantedForest-package.R
\docType{package}
\name{randomPlantedForest-package}
\alias{randomPlantedForest}
\alias{randomPlantedForest-package}
\title{randomPlantedForest: Random Planted Forest: A Directly Interpretable Tree Ensemble}
\description{
An implementation of the Random Planted Forest algorithm for directly interpretable tree ensembles based on a functional ANOVA decomposition.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PlantedML/randomPlantedForest}
  \item \url{http://plantedml.com/randomPlantedForest/}
  \item Report bugs at \url{https://github.com/PlantedML/randomPlantedForest/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{burk@leibniz-bips.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

Authors:
\itemize{
  \item Joseph Theo Meyer
  \item Munir Hiabu
  \item Maike Spankus
  \item Marvin N. Wright \email{cran@wrig.de}
}

}
\keyword{internal}
