% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpf.R
\name{rpf}
\alias{rpf}
\title{Random Planted Forest}
\usage{
rpf(
  Y,
  X,
  max_interaction = 1,
  ntrees = 50,
  splits = 30,
  split_try = 10,
  t_try = 0.4,
  deterministic = FALSE,
  parallel = FALSE,
  purify = FALSE,
  cv = FALSE,
  loss = "L2",
  delta = 0,
  epsilon = 0.1
)
}
\arguments{
\item{Y}{Vector of target features}

\item{X}{Feature matrix}

\item{max_interaction}{Maximum level of interaction determining maximum
number of split dimensions for a tree}

\item{ntrees}{Number of trees generated per family}

\item{splits}{Number of performed splits for each tree family}

\item{split_try}{TODO}

\item{t_try}{TODO}

\item{deterministic}{Choose whether approach deterministic or random}

\item{parallel}{Perform algorithm in parallel or serialized}

\item{purify}{Whether the forest should be purified}

\item{cv}{Determines if cross validation is performed}

\item{loss}{"L2" or "logit" / TODO}

\item{delta}{TODO}

\item{epsilon}{TODO}
}
\value{
Object
}
\description{
Random Planted Forest
}
\examples{
sample_size <- 500
data <- generate_data(Model=1, p=4, n=sample_size)
test_size <- floor(length(data$Y_true) / 5)
x_test <- data$X[1:test_size, ] # extract samples
y_test <- data$Y_true[1:test_size]
x_train <- data$X[(test_size+1):sample_size, ] # extract samples
y_train <- data$Y_start[(test_size+1):sample_size]


# set parameters ------------------------
n_splits <- 15
max_inter <- 2
n_trees <- 50
split_try <- 10
t_try <- 0.5
deterministic_forest <- TRUE
parallel <- TRUE
purify_forest <- FALSE
loss <- 'logit'
delta <- 0.1
epsilon <- 0

# train models ------------------------
rpf_fit <- rpf(
  y_train, x_train,  max_interaction=max_inter, t_try=t_try,
  ntrees=n_trees, splits=n_splits, split_try = split_try,
  deterministic=deterministic_forest, parallel=parallel
)
}
