% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpf.R
\name{rpf}
\alias{rpf}
\alias{rpf.default}
\alias{rpf.data.frame}
\alias{rpf.matrix}
\alias{rpf.formula}
\alias{rpf.recipe}
\title{Random Planted Forest}
\usage{
rpf(x, ...)

\method{rpf}{default}(x, ...)

\method{rpf}{data.frame}(
  x,
  y,
  max_interaction = 1,
  ntrees = 50,
  splits = 30,
  split_try = 10,
  t_try = 0.4,
  deterministic = FALSE,
  n_threads = 1,
  purify = FALSE,
  cv = FALSE,
  loss = "L2",
  delta = 0,
  epsilon = 0.1,
  ...
)

\method{rpf}{matrix}(
  x,
  y,
  max_interaction = 1,
  ntrees = 50,
  splits = 30,
  split_try = 10,
  t_try = 0.4,
  deterministic = FALSE,
  n_threads = 1,
  purify = FALSE,
  cv = FALSE,
  loss = "L2",
  delta = 0,
  epsilon = 0.1,
  ...
)

\method{rpf}{formula}(
  formula,
  data,
  max_interaction = 1,
  ntrees = 50,
  splits = 30,
  split_try = 10,
  t_try = 0.4,
  deterministic = FALSE,
  n_threads = 1,
  purify = FALSE,
  cv = FALSE,
  loss = "L2",
  delta = 0,
  epsilon = 0.1,
  ...
)

\method{rpf}{recipe}(
  x,
  data,
  max_interaction = 1,
  ntrees = 50,
  splits = 30,
  split_try = 10,
  t_try = 0.4,
  deterministic = FALSE,
  n_threads = 1,
  purify = FALSE,
  cv = FALSE,
  loss = "L2",
  delta = 0,
  epsilon = 0.1,
  ...
)
}
\arguments{
\item{x}{Feature matrix or \code{data.frame}.}

\item{...}{(Unused).}

\item{y}{Target vector for use with \code{x}.
The class of \code{y} (either \code{numeric} or \code{\link{factor}}) determines if regression
or classification will be performed.}

\item{max_interaction}{\verb{[1]}: Maximum level of interaction determining maximum
number of split dimensions for a tree.
The default \code{1} corresponds to main effects only.
If \code{0}, the number fo columns in \code{x} is used, i.e. for 10 predictors,
this is equivalent to setting \code{max_interaction = 10}.}

\item{ntrees}{\verb{[50]}: Number of trees generated per family.}

\item{splits}{\verb{[30]}: Number of splits performed for each tree family.}

\item{split_try}{\verb{[10]}: Number of split points to be considered when choosing a split candidate.}

\item{t_try}{\verb{[0.4]}: A value in (0,1] specifying the proportion of viable split-candidates in each round.}

\item{deterministic}{\verb{[FALSE]}: Choose whether approach deterministic or random.}

\item{n_threads}{\verb{[1L]}: Number of threads used for ocmputation, defaulting to \code{1}, i.e. serial execution.}

\item{purify}{\verb{[FALSE]}: Whether the forest should be purified.
Set to \code{TRUE} to enable components extract with \code{\link[=extract_components]{extract_components()}} are valid.
Can be achieved after fitting with \code{\link[=purify]{purify()}}.}

\item{cv}{\verb{[FALSE]}: Determines if cross validation is performed.}

\item{loss}{\verb{["L2"]}: For regression, only \code{"L2"} is supported. For
classification, \code{"L1"}, \code{"logit"} and "\code{exponential}" are also available.
"\code{exponential}" yield similar results as "\code{logit}" while being significantly faster.}

\item{delta}{\verb{[0]}: Only used if loss = \code{"logit"} or \code{"exponential"}.
Proportion of class membership is truncated to be smaller 1-delta when calculating
the loss to determine the optimal split.}

\item{epsilon}{\verb{[0.1]}: Only used if loss = \code{"logit"} or \code{"exponential"}.
Proportion of class membership is truncated to be smaller 1-epsilon when calculating
the fit in a leaf.}

\item{formula}{Formula specification, e.g. y ~ x1 + x2.}

\item{data}{A \code{data.frame} for use with \code{formula}.}
}
\value{
Object of class \code{"rpf"} with model object contained in \verb{$fit}.
}
\description{
Random Planted Forest
}
\examples{
# Regression with x and y
rpfit <- rpf(x = mtcars[, c("cyl", "wt")], y = mtcars$mpg)

# Regression with formula
rpfit <- rpf(mpg ~ cyl + wt, data = mtcars)
}
