% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpf.R
\name{rpf}
\alias{rpf}
\title{Random Planted Forest}
\usage{
rpf(
  Y,
  X,
  max_interaction = 1,
  ntrees = 50,
  splits = 30,
  split_try = 10,
  t_try = 0.4,
  deterministic = FALSE,
  parallel = FALSE,
  purify = FALSE,
  cv = FALSE,
  loss = "L2",
  delta = 0,
  epsilon = 0.1
)
}
\arguments{
\item{Y}{Vector of target features}

\item{X}{Feature matrix}

\item{max_interaction}{Maximum level of interaction determining maximum
number of split dimensions for a tree}

\item{ntrees}{Number of trees generated per family}

\item{splits}{Number of performed splits for each tree family}

\item{split_try}{TODO}

\item{t_try}{TODO}

\item{deterministic}{Choose whether approach deterministic or random}

\item{parallel}{Perform algorithm in parallel or serialized}

\item{purify}{Whether the forest should be purified}

\item{cv}{Determines if cross validation is performed}

\item{loss}{"L2" or "logit" / TODO}

\item{delta}{TODO}

\item{epsilon}{TODO}
}
\value{
Object
}
\description{
Random Planted Forest
}
\examples{
\dontrun{
# TODO (see tests)
}
}
