% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.R
\name{rpf_is_valid}
\alias{rpf_is_valid}
\title{Check whether an rpf object's C++ forest is still alive}
\usage{
rpf_is_valid(x)
}
\arguments{
\item{x}{An object of class \code{rpf}.}
}
\value{
\code{TRUE} if the underlying model can be used, \code{FALSE} otherwise.
}
\description{
The C++ forest does not survive \code{\link[=saveRDS]{saveRDS()}}; an \code{rpf} object restored via
\code{\link[=readRDS]{readRDS()}} without \code{\link[=rpf_marshal]{rpf_marshal()}}/\code{\link[=rpf_unmarshal]{rpf_unmarshal()}} is unusable.
}
